refactor integration test setup
