+++ integration-tests/auth/next.config.js
@@ -4,7 +4,4 @@ const withBundleAnalyzer = require("@next/bundle-analyzer")({
 
 module.exports = withBundleAnalyzer({
   reactStrictMode: true,
-  experimental: {
-    esmExternals: "loose",
-  },
 })
