+++ integration-tests/auth/package.json
@@ -31,19 +31,11 @@
     "@types/get-port": "4.2.0",
     "@types/node-fetch": "2.6.1",
     "@types/react": "17.0.43",
-    "@types/rimraf": "3.0.2",
-    "@types/selenium-webdriver": "4.0.18",
     "b64-lite": "1.4.0",
-    "chromedriver": "100.0.0",
-    "cross-spawn": "7.0.3",
     "eslint": "7.32.0",
-    "express": "4.17.3",
     "fs-extra": "10.0.1",
     "get-port": "6.1.2",
     "node-fetch": "3.2.3",
-    "rimraf": "3.0.2",
-    "selenium-webdriver": "4.1.1",
-    "tree-kill": "1.2.2",
     "typescript": "^4.5.3"
   }
 }
